Major refactor. cable libraries were too buggy to use reliably. This has been reverted to a user adding their own cables for use in the program.
The same functionality has been implemented in the heatshrink and cable braid libraries. These are user accessible files that are saved next to the GUI application for easy access and sharing between computers. It is expected that indivudual users will add and maintain their own cables.

Added ability to rotate layers individually of eachother to verify cable positions